Added popup and options page (not functional).
Adds a popup and options pages with html, css, and
js files for each. These are are accessible through
the browser action but do not have any meaningful
content within them yet, and are placeholders for
now.

The storage API has been added to the
manifest.json, and will be used to save options,
and the list of sites that the focus alert is to
be enabled on.